test: 조회 및 좋아요 기능 테스트 코드 분리
